Smaller Count And Threshold
As previously discussed, and to match the official FROST
implementation from the ZCash Foundation, `count` and `threshold` are
now `uint16`s. This unfortunately added quite a bit of small changes
as `viem` maps `uint64`s to `bigint`s but `uint16`s to `number`s.

There are a couple of other small changes that I did while I was at
it, in particular I made all the `KeyGen` events also include a
"completed" flag for each phase ("committed", "shared", "compromised",
and "confirmed").

Futhermore, I fixed a small issue with the event listener which was
causing complaint flow events to not be processed (as they were
missing from the contract ABI).

+++ contracts/test/FROSTCoordinator.t.sol
@@ -24,8 +24,8 @@ contract FROSTCoordinatorTest is Test {
         ForgeSecp256k1.P e;
     }
 
-    uint64 public constant COUNT = 5;
-    uint64 public constant THRESHOLD = 3;
+    uint16 public constant COUNT = 5;
+    uint16 public constant THRESHOLD = 3;
 
     FROSTCoordinator public coordinator;
     ParticipantMerkleTree public participants;
@@ -365,7 +365,7 @@ contract FROSTCoordinatorTest is Test {
         FROST.verify(groupKey, signature, message);
     }
 
-    function _randomSortedAddresses(uint64 count) private view returns (address[] memory result) {
+    function _randomSortedAddresses(uint16 count) private view returns (address[] memory result) {
         result = new address[](count);
         for (uint256 i = 0; i < result.length; i++) {
             result[i] = vm.randomAddress();
